+++ packages/pull-quote/__tests__/ios/jest.config.js
@@ -1,3 +1,3 @@
 const jestConfigurator = require("@times-components/jest-configurator");
 
-module.exports = jestConfigurator("pull-quotes", "ios");
+module.exports = jestConfigurator("pull-quote", "ios");