+++ packages/pull-quote/__tests__/web/jest.config.js
@@ -1,6 +1,6 @@
 const jestConfigurator = require("@times-components/jest-configurator");
 
-module.exports = Object.assign(jestConfigurator("pull-quotes", "web"), {
+module.exports = Object.assign(jestConfigurator("pull-quote", "web"), {
   collectCoverageFrom: [
     "**/packages/pull-quote/*.js",
     "!**/packages/pull-quote/*stories.js",