+++ packages/article-flag/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.9.0"></a>
+# [0.9.0](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.8.5...@times-components/article-flag@0.9.0) (2017-12-12)
+
+
+### Features
+
+* refactor storybook package ([#453](https://github.com/newsuk/times-components/issues/453)) ([acadef3](https://github.com/newsuk/times-components/commit/acadef3))
+
+
+
+
+<a name="0.8.5"></a>
+## [0.8.5](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.8.4...@times-components/article-flag@0.8.5) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
+<a name="0.8.4"></a>
+## [0.8.4](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.8.3...@times-components/article-flag@0.8.4) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
 <a name="0.8.3"></a>
 ## [0.8.3](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.8.1...@times-components/article-flag@0.8.3) (2017-12-04)