+++ packages/article-label/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.0"></a>
+# [0.6.0](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.5.5...@times-components/article-label@0.6.0) (2017-12-12)
+
+
+### Features
+
+* refactor storybook package ([#453](https://github.com/newsuk/times-components/issues/453)) ([acadef3](https://github.com/newsuk/times-components/commit/acadef3))
+
+
+
+
+<a name="0.5.5"></a>
+## [0.5.5](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.5.4...@times-components/article-label@0.5.5) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-label
+
+<a name="0.5.4"></a>
+## [0.5.4](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.5.3...@times-components/article-label@0.5.4) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-label
+
 <a name="0.5.3"></a>
 ## [0.5.3](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.5.2...@times-components/article-label@0.5.3) (2017-12-05)