+++ packages/watermark/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.1.4...@times-components/watermark@0.2.0) (2017-12-12)
+
+
+### Features
+
+* refactor storybook package ([#453](https://github.com/newsuk/times-components/issues/453)) ([acadef3](https://github.com/newsuk/times-components/commit/acadef3))
+
+
+
+
+<a name="0.1.4"></a>
+## [0.1.4](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.1.3...@times-components/watermark@0.1.4) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
+<a name="0.1.3"></a>
+## [0.1.3](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.1.2...@times-components/watermark@0.1.3) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
 <a name="0.1.2"></a>
 ## [0.1.2](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.1.0...@times-components/watermark@0.1.2) (2017-12-04)